Add wake lock to proximity sensor.
This shouldn't really be needed, but will solve the problem of the device sleeping in the middle of a call. It has potential to drain battery if an app doesn't disable the proximity sensor when closing, but I suspect that was happening anyway if the device went to sleep while the sensor was on. Apps should always disable sensors when no longer needed, so this should not cause any problems.