msm_fb: display: Fix for HDMI bootup issue
If the device is booted up with HDMI cable connected, HDTV would go
blank unless the HDMI cable is removed and inserted again. This is
happening because the dtv pipe is initialized after HDMI core is
initialized. The dtv pipe needs to be allocated at the bootup time
when dtv is turned on to syncronize the HDMI core and DTV overlay.

CRs-fixed: 331867
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
(cherry picked from commit a369ce03ca048fb511642ac1cb73c1f652d70c50)

Change-Id: I1dbb21e393919c5aaca4b0a07d02de42f4cb6223
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>